requirements.txt
requests== 2.32.3
matplotlib==3.9.1
numpy==1.26.3
scikit-learn==1.5.1 
transformers==4.43.2
rake-nltk==1.0.6
torch==2.4.0+cu118
langchain==0.2.11
langchain-community==0.2.10
langchain-text-splitters==0.2.2
langchain-core==0.2.23
streamlit==1.36.0
PyMuPDF==1.24.9
reportlab==4.2.2
gTTS==2.5.2
deep-translator==1.11.4
pillow==10.2.0
python-docx==1.1.2
nltk==3.8.1
docx2pdf==0.1.8
fitz==0.0.1.dev2


